Fixed issue with form not validating correctly and rearranged how emails are sent
Form was not validating because is_valid() used the default value from the kwargs.pop() as the max. NOt sure why, but just raised the default value and it's working.  Also will only send the user an email if they accept the reservation.  I never get their email if the don't, so impossible to send them an email...